Move `iter_submodules()` to `iter_collections
It is practically a thin wrapper around `iter_gitworktree()` and
the set of dependencies is therefore identical. Spreading the
two across two packages is an invitation for circular dependencies.

Moreover, there are no dedicated tests in `repo_utils`.
