+++ datalad_next/iter_collections/gitworktree.py
@@ -235,6 +235,28 @@ def iter_gitworktree(
         # report in the next loop iteration
 
 
+def iter_submodules(
+    path: Path,
+) -> Generator[GitTreeItem, None, None]:
+    """Given a path, report all submodules of a repository worktree underneath
+
+    This is a thin convenience wrapper around ``iter_gitworktree()``.
+    """
+    for item in iter_gitworktree(
+        path,
+        untracked=None,
+        link_target=False,
+        fp=False,
+        recursive='repository',
+    ):
+        # exclude non-submodules, or a submodule that was found at
+        # the root path -- which would indicate that the submodule
+        # itself it not around, only its record in the parent
+        if item.gittype == GitTreeItemType.submodule \
+                and item.name != PurePath('.'):
+            yield item
+
+
 def _get_item(
     basepath: Path,
     link_target: bool,
